Laravel

+++ app/Http/Controllers/PostsController.php
@@ -3,8 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
+use Image;
+use App\Country;
 
 class PostsController extends Controller
 {
@@ -29,11 +32,12 @@ class PostsController extends Controller
     public function index()
     {
         //
-       //$posts= Post::all();
+        //$posts= Post::all();
         //$posts= post::orderBy('title','asc')->take(1)->get(); 1 refer to the first post
-      // $posts= post::orderBy('title','asc')->get();
-        $posts= post::orderBy('created_at','asc')->paginate(10); //make pages number to scrol
+        // $posts= post::orderBy('title','asc')->get();
+        $posts= post::all()->paginate(10); //make pages number to scrol
         return view('posts.index')->with ('posts',$posts);
+
     }
 
     /**
@@ -55,18 +59,40 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-     $this->validate($request,[
-         'title'=>'required',
-         'body'=>'required',
-     ])   ;
-     //create post
+        $this->validate($request,[
+            'title'=>'required',
+            'body'=>'required',
+            'country'=>'required',
+            'cover_image' => 'image|nullable|max:1999'
+        ]);
+        //handle file
+        if($request->hasFile('cover_image')){
+            $file = $request->file('cover_image');
+
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            Image::make($file)->insert(storage_path("app/public/cover_images/watermark.png"))->save(storage_path("app/public/cover_images/" . $fileNameToStore));
+        } else {
+            $fileNameToStore = 'noImage.jpg';
+        }
+
+        //create post
         $post=new Post;
         $post->title=$request->input('title');
         $post->body=$request->input('body');
         $post->user_id =auth()->user()->id;//look to current user
+        $post->cover_image = $fileNameToStore;
+        $post->country_id = $request->input('country') + 1;
         $post->save();
 
-        return redirect('posts')->with ('success','post created');
+        return redirect('/')->with ('success','post created');
     }
 
     /**
@@ -82,9 +108,9 @@ class PostsController extends Controller
 
 
         //return=Post::where('title','Post Two')->get();
-       // $posts=DB::select('select * from posts');
+        // $posts=DB::select('select * from posts');
         //$posts= Post::orderBy('title','desc')->get();
-       // $posts= post::orderBy('title','desc')->get();
+        // $posts= post::orderBy('title','desc')->get();
         $post= Post::find($id);
         return view('posts.show')->with('post',$post);
     }
@@ -114,15 +140,37 @@ class PostsController extends Controller
         $this->validate($request,[
             'title'=>'required',
             'body'=>'required',
-        ])   ;
-        //create post
+            'country'=>'required',
+            'cover_image' => 'image|nullable|max:1999'
+        ]);
         $post=Post::find($id);
+        //handle file
+        if($request->hasFile('cover_image')){
+            $file = $request->file('cover_image');
+
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            Image::make($file)->insert(storage_path("app/public/cover_images/watermark.png"))->save(storage_path("app/public/cover_images/" . $fileNameToStore));
+        } else {
+            $fileNameToStore = $post->cover_image;
+        }
+
+        //create post
         $post->title=$request->input('title');
         $post->body=$request->input('body');
-       // $post->user_id =auth()->user()->id;//look to current user
+        $post->user_id =auth()->user()->id;//look to current user
+        $post->cover_image = $fileNameToStore;
+        $post->country_id = $request->input('country') + 1;
         $post->save();
 
-        return redirect('posts')->with ('success','Post Updated');
+        return redirect('/')->with ('success','post created');
     }
 
     /**
@@ -133,9 +181,18 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        $post=Post::find($id);
+        $post = Post::find($id);
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id && auth()->user()->admin!==1){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        if($post->cover_image != 'noImage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
-        return redirect('posts')->with('success','Post Removed');
+        return redirect('/dashboard')->with('success', 'Post Removed');
 
     }
 }
